ow-actions-cdk/requirements.txt
aws-cdk.core>=1.132.0
aws-cdk.assertions>=1.132.0
aws-cdk.aws_sns>=1.132.0
aws_cdk.aws_sns_subscriptions>=1.132.0
